Moved simulation code into a dedicated module
