+++ backend/counterfactual/__init__.py
@@ -1,0 +1,1 @@
+from .simulation import simulate_records
